Extract Keycloak Context out of bootstrap
