+++ src/KeycloakContext.ts
@@ -1,0 +1,4 @@
+import { createContext } from 'react'
+import Keycloak from 'keycloak-js'
+
+export const KeycloakContext = createContext<Keycloak>(new Keycloak())
