+++ src/components/Login/LoginButton.tsx
@@ -3,7 +3,7 @@ import { deepOrange } from '@mui/material/colors'
 import Keycloak, { KeycloakTokenParsed } from 'keycloak-js'
 import { ReactElement, useContext, useState } from 'react'
 import { LoginPanel } from './LoginPanel'
-import { KeycloakContext } from '../../bootstrap'
+import { KeycloakContext } from '../../KeycloakContext'
 import { AppConfigContext } from '../../AppConfigContext'
 import { logApi } from '../../debug'
