+++ src/components/ToolBar/DataMenu/LoadFromNdexDialog.tsx
@@ -30,7 +30,7 @@ import { NDEx } from '@js4cytoscape/ndex-client'
 import { useWorkspaceStore } from '../../../store/WorkspaceStore'
 import { ndexSummaryFetcher } from '../../../store/getNetworkSummaryFromCacheOrNdex'
 import { dateFormatter } from '../../../utils/date-format'
-import { KeycloakContext } from '../../../bootstrap'
+import { KeycloakContext } from '../../../KeycloakContext'
 import { useMessageStore } from '../../../store/MessageStore'
 import { MessageSeverity } from '../../../models/MessageModel'
 import { logUi } from '../../../debug'
