+++ src/components/ToolBar/DataMenu/SaveToNDExMenuItem.tsx
@@ -25,7 +25,7 @@ import { useNetworkSummaryStore } from '../../../store/NetworkSummaryStore'
 import { Network } from '../../../models/NetworkModel'
 import { AppConfigContext } from '../../../AppConfigContext'
 import { useMessageStore } from '../../../store/MessageStore'
-import { KeycloakContext } from '../../../bootstrap'
+import { KeycloakContext } from '../../../KeycloakContext'
 import { useHcxValidatorStore } from '../../../features/HierarchyViewer/store/HcxValidatorStore'
 import { HcxValidationSaveDialog } from '../../../features/HierarchyViewer/components/Validation/HcxValidationSaveDialog'
 import { NetworkView } from '../../../models/ViewModel'
