+++ src/components/ToolBar/DataMenu/SaveWorkspaceToNDExOverwrite.tsx
@@ -6,7 +6,7 @@ import { NDEx } from '@js4cytoscape/ndex-client'
 import { useCredentialStore } from '../../../store/CredentialStore'
 import { AppConfigContext } from '../../../AppConfigContext'
 import { useMessageStore } from '../../../store/MessageStore'
-import { KeycloakContext } from '../../../bootstrap'
+import { KeycloakContext } from '../../../KeycloakContext'
 import { useSaveWorkspace } from '../../../utils/ndex-utils'
 import { ConfirmationDialog } from '../../Util/ConfirmationDialog'
 import { MessageSeverity } from '../../../models/MessageModel'
